@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -36,8 +36,10 @@ dependencies:
   cupertino_icons: ^1.0.0
   english_words: ^4.0.0
   provider: ^6.0.3
-  flutter_staggered_grid_view: 0.4.0
+  flutter_staggered_grid_view: ^0.4.0
   flutter_svg: ^1.0.3
+  google_fonts: ^3.0.1
+
 
 
 dev_dependencies:
