@@ -3,7 +3,8 @@ import 'package:grodeals/common_widgets/app_text.dart';
 import 'package:grodeals/models/category_item.dart';
 
 class CategoryItemCardWidget extends StatelessWidget {
-  CategoryItemCardWidget({Key key, this.item, this.color = Colors.blue})
+  const CategoryItemCardWidget(
+      {Key? key, required this.item, this.color = Colors.blue})
       : super(key: key);
   final CategoryItem item;
 
@@ -11,14 +12,14 @@ class CategoryItemCardWidget extends StatelessWidget {
 
   final width = 175.0;
 
-  final Color borderColor = Color(0xffE2E2E2);
+  final Color borderColor = const Color(0xffE2E2E2);
   final double borderRadius = 18;
   final Color color;
 
   @override
   Widget build(BuildContext context) {
     return Container(
-      padding: EdgeInsets.only(top: 5, bottom: 5),
+      padding: const EdgeInsets.only(top: 5, bottom: 5),
       decoration: BoxDecoration(
         color: color.withOpacity(0.1),
         borderRadius: BorderRadius.circular(18),
@@ -30,7 +31,7 @@ class CategoryItemCardWidget extends StatelessWidget {
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.center,
         children: [
-          Container(
+          SizedBox(
             height: 120,
             width: 120,
             child: imageWidget(),
@@ -52,11 +53,9 @@ class CategoryItemCardWidget extends StatelessWidget {
   }
 
   Widget imageWidget() {
-    return Container(
-      child: Image.asset(
-        item.imagePath,
-        fit: BoxFit.contain,
-      ),
+    return Image.asset(
+      item.imagePath,
+      fit: BoxFit.contain,
     );
   }
 }
