@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
 import 'package:grodeals/common_widgets/app_text.dart';
 import 'package:grodeals/models/category_item.dart';
@@ -21,22 +20,24 @@ List<Color> gridColors = [
 ];
 
 class ExploreScreen extends StatelessWidget {
+  const ExploreScreen({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: const CustomAppBar(
+        appBar: CustomAppBar(
           height: 50,
         ),
         body: SafeArea(
-      child: Column(
-        children: [
-          getHeader(),
-          Expanded(
-            child: getStaggeredGridView(context),
+          child: Column(
+            children: [
+              getHeader(),
+              Expanded(
+                child: getStaggeredGridView(context),
+              ),
+            ],
           ),
-        ],
-      ),
-    ));
+        ));
   }
 
   Widget getHeader() {
@@ -47,6 +48,7 @@ class ExploreScreen extends StatelessWidget {
         ),
         Center(
           child: AppText(
+            key: UniqueKey(),
             text: "Find Products",
             fontSize: 20,
             fontWeight: FontWeight.bold,
@@ -66,6 +68,13 @@ class ExploreScreen extends StatelessWidget {
   Widget getStaggeredGridView(BuildContext context) {
     return StaggeredGridView.count(
       crossAxisCount: 4,
+
+      //Here is the place that we are getting flexible/ dynamic card for various images
+      staggeredTiles: categoryItemsDemo
+          .map<StaggeredTile>((_) => StaggeredTile.fit(2))
+          .toList(),
+      mainAxisSpacing: 3.0,
+      crossAxisSpacing: 4.0,
       children: categoryItemsDemo.asMap().entries.map<Widget>((e) {
         int index = e.key;
         CategoryItem categoryItem = e.value;
@@ -74,28 +83,21 @@ class ExploreScreen extends StatelessWidget {
             onCategoryItemClicked(context, categoryItem);
           },
           child: Container(
-            padding: EdgeInsets.all(10),
+            padding: const EdgeInsets.all(10),
             child: CategoryItemCardWidget(
               item: categoryItem,
               color: gridColors[index % gridColors.length],
             ),
           ),
         );
-      }).toList(),
-
-      //Here is the place that we are getting flexible/ dynamic card for various images
-      staggeredTiles: categoryItemsDemo
-          .map<StaggeredTile>((_) => StaggeredTile.fit(2))
-          .toList(),
-      mainAxisSpacing: 3.0,
-      crossAxisSpacing: 4.0, // add some space
+      }).toList(), // add some space
     );
   }
 
   void onCategoryItemClicked(BuildContext context, CategoryItem categoryItem) {
-    Navigator.of(context).push(new MaterialPageRoute(
+    Navigator.of(context).push(MaterialPageRoute(
       builder: (BuildContext context) {
-        return CategoryItemsScreen();
+        return const CategoryItemsScreen();
       },
     ));
   }
